third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/shadow-dom-modify-chardata-expected.txt

Tests that elements panel updates shadow dom tree structure upon typing.



Running: testDumpInitial
========= Original ========
- <div id="container">
    - <input type="text" id="input1">
        - #shadow-root (user-agent)
              <div id="inner-editor"></div>
      </input>
  </div>

Running: testAppend
======== Type text =========
- <div id="container">
    - <input type="text" id="input1">
        - #shadow-root (user-agent)
              <div id="inner-editor">Bar</div>
      </input>
  </div>
